fix BEO-4782

+++ DAL/CatalogDAL.cs
@@ -5407,7 +5407,7 @@ namespace Tvinci.Core.DAL
         }
 
         public static DataSet InsertChannel(int groupId, string systemName, string name, string description, int isActive, int orderBy, int orderByDir, string orderByValue, int channelType, string filterQuery, List<int> assetTypes, string groupBy,
-                                            List<KeyValuePair<string, string>> namesInOtherLanguages, List<KeyValuePair<string, string>> descriptionsInOtherLanguages, List<KeyValuePair<int, int>> mediaIdsToOrderNum, long userId)
+                                            List<KeyValuePair<string, string>> namesInOtherLanguages, List<KeyValuePair<string, string>> descriptionsInOtherLanguages, List<KeyValuePair<long, int>> mediaIdsToOrderNum, long userId)
         {
             ODBCWrapper.StoredProcedure sp = new ODBCWrapper.StoredProcedure("InsertChannel");
             sp.SetConnectionKey("MAIN_CONNECTION_STRING");
@@ -5429,14 +5429,14 @@ namespace Tvinci.Core.DAL
             sp.AddParameter("@DescriptionsInOtherLanguagesExist", descriptionsInOtherLanguages != null && descriptionsInOtherLanguages.Count > 0);
             sp.AddKeyValueListParameter<string, string>("@DescriptionsInOtherLanguages", descriptionsInOtherLanguages, "key", "value");
             sp.AddParameter("@MediaIdsToOrderNumExist", mediaIdsToOrderNum != null && mediaIdsToOrderNum.Count > 0 ? 1 : 0);
-            sp.AddKeyValueListParameter<int, int>("@MediaIdsToOrderNum", mediaIdsToOrderNum, "key", "value");
+            sp.AddKeyValueListParameter<long, int>("@MediaIdsToOrderNum", mediaIdsToOrderNum, "key", "value");
             sp.AddParameter("@UpdaterID", userId);
 
             return sp.ExecuteDataSet();
         }
 
         public static DataSet UpdateChannel(int groupId, int id, string systemName, string name, string description, int isActive, int orderBy, int orderByDir, string orderByValue, string filterQuery, List<int> assetTypes, string groupBy,
-            List<KeyValuePair<string, string>> namesInOtherLanguages, List<KeyValuePair<string, string>> descriptionsInOtherLanguages, List<KeyValuePair<int, int>> mediaIdsToOrderNum, long userId)
+            List<KeyValuePair<string, string>> namesInOtherLanguages, List<KeyValuePair<string, string>> descriptionsInOtherLanguages, List<KeyValuePair<long, int>> mediaIdsToOrderNum, long userId)
         {
             ODBCWrapper.StoredProcedure sp = new ODBCWrapper.StoredProcedure("UpdateChannel");
             sp.SetConnectionKey("MAIN_CONNECTION_STRING");
@@ -5458,7 +5458,7 @@ namespace Tvinci.Core.DAL
             sp.AddParameter("@ShouldUpdateOtherDescriptions", descriptionsInOtherLanguages != null && descriptionsInOtherLanguages.Count > 0);
             sp.AddKeyValueListParameter<string, string>("@DescriptionsInOtherLanguages", descriptionsInOtherLanguages, "key", "value");
             sp.AddParameter("@MediaIdsToOrderNumExist", mediaIdsToOrderNum != null && mediaIdsToOrderNum.Count > 0 ? 1 : 0);
-            sp.AddKeyValueListParameter<int, int>("@MediaIdsToOrderNum", mediaIdsToOrderNum, "key", "value");
+            sp.AddKeyValueListParameter<long, int>("@MediaIdsToOrderNum", mediaIdsToOrderNum, "key", "value");
             sp.AddParameter("@UpdaterID", userId);
 
             return sp.ExecuteDataSet();
